+++ CHANGELOG.md
@@ -8,6 +8,9 @@ All notable changes to this project will be documented in this file.
 - Choose type of client on creation.
 - Implement Proof Key for Code Exchange by OAuth Public Clients.
 
+##### Fixed
+- Not auto-approve requests for non-confidential clients (publics).
+
 ### [0.3.1] - 2016-03-09
 
 ##### Fixed