+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to this project will be documented in this file.
 - Choose type of client on creation.
 - Implement Proof Key for Code Exchange by OAuth Public Clients.
 - Support for prompt parameter.
+- Support for different client JWT tokens algorithm.
 
 ##### Fixed
 - Not auto-approve requests for non-confidential clients (publics).